acceptor: split acceptor.c
Currently, changes between TCP and UDS is inter-mixed, so split them up into
their own files. This is cleaner and is needed to support HTTP/3 in the future.
